+++ src/test/java/ru/yandex/practicum/filmorate/controller/userTests/UserControllerTest1.java
@@ -1,8 +1,9 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.controller.userTests;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import ru.yandex.practicum.filmorate.controller.UserController;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.time.LocalDate;
@@ -11,7 +12,7 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SpringBootTest
-class ShouldCreateUserInUserControllerTest1 {
+class UserControllerTest1 {
     @Autowired
     private UserController userController;
     User user;
@@ -28,7 +29,7 @@ class ShouldCreateUserInUserControllerTest1 {
         userController.addUser(user);
         List<User> allUsers = userController.getAllUser();
 
-        assertEquals(1, allUsers.size(), "Size Equal Test");
+        assertEquals(4, allUsers.size(), "Size Equal Test");
 
         User userToCheck = allUsers.get(0);
