+++ src/test/java/ru/yandex/practicum/filmorate/controller/userTests/UserControllerTest2.java
@@ -1,43 +1,39 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.controller.userTests;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.controller.UserController;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.time.LocalDate;
 import java.util.List;
 
-import static org.junit.Assert.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
-class ShouldGetAllUserTestInUserControllerTest4 {
+class UserControllerTest2 {
     @Autowired
     private UserController userController;
     User user;
 
     @Test
-    public void shouldGetAllUserTest() {
+    public void shouldCreateUserWithoutName() {
         user = User.builder()
                 .login("dolore")
                 .name("Nick Name")
                 .email("mail@mail.ru")
                 .birthday(LocalDate.of(1946, 8, 20))
                 .build();
+        user.setName("");
         userController.addUser(user);
-        User user100 = user;
-        user100.setId(100);
 
-        userController.addUser(user100);
+        List<User> allUsers = userController.getAllUser();
 
-        assertEquals(2, userController.getAllUser().size(),
-                "Size Equal Test");
+        User userToCheck = allUsers.get(1);
+
+        assertEquals(user.getLogin(), userToCheck.getName(), "Login Equal Name Test");
     }
+
+
 }
